// src/interfaces/location.ts
export interface LocationResult {
  country?: string;
  city?: string;
  lng?: number;
  searchType?: string;
  alternative?: string[];
  index?: number;
  bookingId?: string;
  placeType?: string;
  placeKey?: string;
  countryIso?: string;
  locationId?: string;
  name?: string;
  region?: string;
  ufi?: number;
  isPopular?: boolean;
  lang?: string;
  lat?: number;
  iata?: string;
}